new Stat system, new EffectInfos

+++ Assets/Project/Scripts/BuffSystem/Buffs/OnFire.cs
@@ -1,6 +1,5 @@
 ﻿using Project.Scripts.BuffSystem.Components;
 using Project.Scripts.BuffSystem.Data;
-using Project.Scripts.EffectSystem.Effects;
 using Project.Scripts.EffectSystem.Effects.Attacks;
 using UnityEngine;
 
@@ -8,7 +7,7 @@ namespace Project.Scripts.BuffSystem.Buffs
 {
     public class OnFire : TickingBuff<IDamageable>
     {
-        public OnFire(float damage, float duration, float tickInterval, IDamageable target, GameObject source)
+        public OnFire(int damage, float duration, float tickInterval, IDamageable target, GameObject source)
             : base(new FireAttack(source, damage), duration, StackBehavior.Stacking, target, tickInterval)
         {
         }
